refactor: optimize the code of the server part
