+++ server/src/controllers/linkControllers.ts
@@ -1,11 +1,8 @@
 import { NextFunction, Request, Response } from "express";
-import IP from "ip";
-import geoip from "geoip-lite";
-import browser from "browser-detect";
 import Link from "../models/link";
 import Statistic from "../models/statistic";
 import ApiError from "../error/apiError";
-import { createDate, createUrl, tryCatchWrapper } from "../libs/server";
+import { createStatistic, createUrl, tryCatchWrapper } from "../libs/server";
 import { ERROR_NOT_FOUND } from "../libs/constants";
 
 class LinkControllers {
@@ -48,23 +45,7 @@ class LinkControllers {
         const { id } = req.params;
         const link = await Link.findByPk(id);
         if (link) {
-          const ipAddress = IP.address(); // ip user
-          const geo = geoip.lookup(ipAddress);
-          // const region = geo.country; // region user
-          const userInf = browser(req.headers["user-agent"]);
-          const userBrowserName = userInf.name; // name browser
-          const userBrowserVersion = userInf.version; // version browser
-          const userOs = userInf.os; // os user
-          const date = createDate(new Date(Date.now())); // request date
-          Statistic.create({
-            data: date,
-            ip: ipAddress,
-            region: "Belarus",
-            browserName: userBrowserName,
-            browserVersion: userBrowserVersion,
-            oc: userOs,
-            LinkId: id,
-          });
+          createStatistic(req, link.id);
           res.redirect(link.originalUrl);
         } else {
           res.sendStatus(404);
