+++ server/src/routes/router.ts
@@ -1,9 +1,7 @@
 import { Router } from "express";
-import statisticRouter from "./statistic";
 import linkRouter from "./link";
 
 const router = Router();
 router.use("/", linkRouter);
-router.use("/statistic", statisticRouter);
 
 export default router;
