--- server/src/routes/statistic.ts
@@ -1,9 +1,0 @@
-import { Router } from "express";
-import statisticControllers from "../controllers/statisticControllers";
-
-const statisticRouter = Router();
-
-statisticRouter.post("/", statisticControllers.createStatistic);
-statisticRouter.get("/", statisticControllers.getStatistic);
-
-export default statisticRouter;
